+++ NEWS.md
@@ -1,5 +1,11 @@
 # connectapi (development version)
 
+## New features
+
+- `get_users()` now supports filtering users with the `account_status` and
+  `user_role` parameters. For example, this allows you to find all licensed
+  users on a Connect server. (#311)
+
 # connectapi 0.4.0
 
 ## New features